+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/SlimefunItem.java
@@ -377,7 +377,7 @@ public class SlimefunItem implements Placeable {
 
             if (SlimefunPlugin.getItemCfg().getBoolean(id + ".enabled")) {
 
-                if (!SlimefunPlugin.getRegistry().getEnabledCategories().contains(category)) {
+                if (!SlimefunPlugin.getRegistry().getCategories().contains(category)) {
                     category.register();
                 }