+++ src/main/java/me/mrCookieSlime/Slimefun/api/Slimefun.java
@@ -8,8 +8,6 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.scheduler.BukkitTask;
 
-import io.github.thebusybiscuit.cscorelib2.config.Config;
-import io.github.thebusybiscuit.slimefun4.api.items.ItemSetting;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
 import io.github.thebusybiscuit.slimefun4.implementation.items.VanillaItem;
 import me.mrCookieSlime.Slimefun.SlimefunPlugin;
@@ -28,36 +26,6 @@ public final class Slimefun {
 
     public static Logger getLogger() {
         return SlimefunPlugin.instance.getLogger();
-    }
-
-    /**
-     * Returns the value associated to this key for the SlimefunItem corresponding to this id.
-     *
-     * @deprecated Please use the {@link ItemSetting} API instead.
-     *
-     * @param id
-     *            the id of the SlimefunItem, not null
-     * @param key
-     *            the key of the value to get, not null
-     *
-     * @return the value associated to the key for the SlimefunItem corresponding to the id,
-     *         or null if it doesn't exist.
-     */
-    @Deprecated
-    public static Object getItemValue(String id, String key) {
-        return getItemConfig().getValue(id + '.' + key);
-    }
-
-    /**
-     * Returns the {@link Config} instance of our Items.yml file.
-     *
-     * @deprecated Do not access this directly, use the {@link ItemSetting} API instead.
-     *
-     * @return the Items.yml Config instance.
-     */
-    @Deprecated
-    public static Config getItemConfig() {
-        return SlimefunPlugin.getItemCfg();
     }
 
     /**
@@ -90,13 +58,7 @@ public final class Slimefun {
     }
 
     public static void registerResearch(NamespacedKey key, int id, String name, int cost, ItemStack... items) {
-        Research research = new Research(key, id, name, cost);
-
-        for (ItemStack item : items) {
-            research.addItems(SlimefunItem.getByItem(item));
-        }
-
-        research.register();
+        registerResearch(new Research(key, id, name, cost), items);
     }
 
     /**
@@ -118,7 +80,10 @@ public final class Slimefun {
 
         if (sfItem != null) {
             if (sfItem.getState() == ItemState.DISABLED) {
-                if (message) SlimefunPlugin.getLocal().sendMessage(p, "messages.disabled-item", true);
+                if (message) {
+                    SlimefunPlugin.getLocal().sendMessage(p, "messages.disabled-item", true);
+                }
+
                 return false;
             }
 
@@ -167,6 +132,7 @@ public final class Slimefun {
                 return false;
             }
         }
+
         return false;
     }